Bug 1171605 - Intern Request Form

+++ extensions/BMO/Extension.pm
@@ -1883,8 +1883,8 @@ sub enter_bug_start {
     # if configured with create_bug_formats, force users into a custom bug
     # format (can be overridden with a __standard__ format)
     my $cgi = Bugzilla->cgi;
-    if ($cgi->param('format') && $cgi->param('format') eq '__standard__') {
-        $cgi->delete('format');
+    if ($cgi->param('format')) {
+        $cgi->delete('format') if $cgi->param('format') eq '__standard__';
     } elsif (my $format = forced_format($cgi->param('product'))) {
         $cgi->param('format', $format);
     }
